User/backend/requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
sentence-transformers==2.7.0
transformers==4.41.2
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
torch==2.2.2
nltk==3.8.1
pymongo[srv]==4.6.1
PyMuPDF==1.24.10
PyPDF2>=3.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.17.0
langchain>=0.3.0
langchain-community>=0.3.0
pypdf>=5.0.0
pytest==8.3.4
PyJWT==2.8.0
bcrypt==4.1.2
